V0.1.0
* Updated base plugin files

Updated addon version from v0.0.0 to v0.1.0

* Game Singleton controller

This singleton provides the data filepaths that game uses at runtime

* Delete YumeVNGlobalVars.gd

Placeholder file - Unused

* Update main plugin files

Updated version info in plugin.cfg file.  Incremented version from 0.0.0 to 0.1.0.

Added an autoload signal that (when called) assigns the current projects game controller singleton to the autoload files.

* New actions and components added

A new set of actions (and accompanying components) were added.

* Updated data saving code

* Updated folder - No changed made

* Added new icons and updated some older ones

* Updated some size flags

* Updated some data

* margin adjustments

* Updated data save process

* Updated data save process

* icon

* Updated data save process

* Updated game components

Updated parser and game components

* font files

* Updated game template items

* Added a yield to allow debug flag change

Without waiting for the debug flag to be set before adding the instance to the scene, the editor's preview scene will crash on first time setup and not be available for use.  It worked on subsequent runs of the editor, but this behavior is not acceptable.

* Updated references to game controller singleton in preview

* Updated references to game controller singleton in parser

* Project filepath validation reintroduced

Reintroduced the need to create the project in an empty directory.  This solves filepaths not initializing correctly.

* Updated debug check

Prevents warning in editor for release mode file check

* Updated debug check